backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.10.4
motor==3.3.1 
pydantic_settings==2.0.0
pydantic[email]==2.5.1
replicate==1.0.4
